+++ envirobot-2.0-module-stm32-software-master/Src/Subscribers/MasterSubscribers.hpp
@@ -48,9 +48,9 @@ public:
 
 	bool SubscribeToRemoteRegister(uint8_t ID, uint8_t register_ID);
 	bool SubscribeToRemoteRegister(uint8_t ID, Register remote_register);
-	bool SubscribeToRemoteRegister(uint8_t ID, uint8_t register_ID, SubscriberInterface interface);
+	bool SubscribeToRemoteRegister(uint8_t ID, uint16_t register_ID, SubscriberInterface interface);
 	bool SubscribeToRemoteRegister(uint8_t ID, Register remote_register, SubscriberInterface interface);
-	bool UnsubscribeFromRemoteRegister(uint8_t ID, uint8_t remote_register_ID);
+	bool UnsubscribeFromRemoteRegister(uint8_t ID, uint16_t remote_register_ID);
 	bool UnsubscribeFromAllRemoteRegisters(uint8_t ID);
 
 	bool AddFilter(uint8_t ID, uint8_t address);
